// src/theme/types/space.ts
export interface Space {
  /** '0px' */
  s000: string;
  /** '2px' */
  s025: string;
  /** '4px' */
  s050: string;
  /** '8px' */
  s100: string;
  /** '16px' */
  s200: string;
  /** '20px' */
  s225: string;
  /** '24px' */
  s250: string;
  /** '28px' */
  s275: string;
  /** '32px' */
  s300: string;
  /** '40px' */
  s325: string;
  /** '48px' */
  s350: string;
  /** '56px' */
  s375: string;
  /** '64px' */
  s400: string;
  /** '128px' */
  s500: string;
  /** '256px' */
  s600: string;
}