// src/theme/types/colors.ts
export type ColorAliasesKeys =
  | 'defaultAlpha64'
  | 'defaultAlpha48'
  | 'defaultAlpha32'
  | 'defaultAlpha20'
  | 'defaultAlpha12'
  | 'defaultAlpha8'
  | 'defaultAlpha4'
  | 'primaryAlpha64'
  | 'primaryAlpha48'
  | 'primaryAlpha32'
  | 'primaryAlpha20'
  | 'primaryAlpha12'
  | 'primaryAlpha8'
  | 'primaryAlpha4'
  | 'accentAlpha64'
  | 'accentAlpha48'
  | 'accentAlpha32'
  | 'accentAlpha20'
  | 'accentAlpha12'
  | 'accentAlpha8'
  | 'accentAlpha4'
  | 'whiteAlpha40'
  | 'whiteAlpha32'
  | 'whiteAlpha24'
  | 'whiteAlpha16'
  | 'whiteAlpha12';

export type ColorKeys =
  | 'default'
  | 'defaultAdd1'
  | 'defaultAdd2'
  | 'defaultAdd3'
  | 'defaultAdd4'
  | 'defaultAdd5'
  | 'defaultAdd6'
  | 'primarySub2'
  | 'primarySub1'
  | 'primary'
  | 'primaryAdd1'
  | 'primaryAdd2'
  | 'primaryAdd3'
  | 'primaryAdd4'
  | 'primaryAdd5'
  | 'primaryAdd6'
  | 'accentSub2'
  | 'accentSub1'
  | 'accent'
  | 'accentAdd1'
  | 'accentAdd2'
  | 'accentAdd3'
  | 'accentAdd4'
  | 'accentAdd5'
  | 'accentAdd6'
  | 'white';

export type Colors = {
  [key in ColorKeys]: string;
} & { [key in ColorAliasesKeys]?: string };